Updated some vcproj files so the ExampleAIModule, ExampleAIClient, AIModuleLoader, and TestAIModule can be built and run in Debug mode (assuming BWAPI is also built in debug mode). This only affects the Debug configuration - Everything can still be built and ran in release as before.
git-svn-id: https://bwapi.googlecode.com/svn/trunk@2470 273ecf2c-f24b-0410-b6a7-e9298bb4873f

+++ bwapi/BWAPI/Source/BWAPI/CommandUpgrade.cpp
@@ -25,7 +25,7 @@ namespace BWAPI
     {
       this->executors[0]->getRawDataLocal()->orderID = BW::OrderID::Upgrade;
       this->executors[0]->getRawDataLocal()->childUnitUnion2.unitIsNotScarabInterceptor.subChildUnitUnion2.upgradeID = upgrade.getID();
-      this->executors[0]->getRawDataLocal()->childUnitUnion1.unitIsBuilding.upgradeResearchTime = upgrade.timeCostBase()+upgrade.timeCostFactor()*this->executors[0]->getPlayer()->getUpgradeLevel(upgrade.getID());
+      this->executors[0]->getRawDataLocal()->childUnitUnion1.unitIsBuilding.upgradeResearchTime = (u16)(upgrade.timeCostBase()+upgrade.timeCostFactor()*this->executors[0]->getPlayer()->getUpgradeLevel(upgrade.getID()));
     }
   }
   //------------------------------------------------ GET TYPE ------------------------------------------------
